+++ Core/Systems/Gameboy/IMBC.cs
@@ -3,9 +3,9 @@ namespace bEmu.Core.Systems
     public interface IMBC
     {
         void LoadProgram(byte[] bytes);
-        byte ReadCartRAM(ushort addr);
-        byte ReadROM(ushort addr);
-        void SetMode(ushort addr, byte value);
-        void WriteCartRAM(ushort addr, byte value);
+        byte ReadCartRAM(int addr);
+        byte ReadROM(int addr);
+        void SetMode(int addr, byte value);
+        void WriteCartRAM(int addr, byte value);
     }
 }
